GlobalTypeOptimization: Handle side effects in removed fields in struct.new (#4263)
If struct.new operands have side effects, and we are removing the operand
as the field is removed, we must keep the side effects. To handle that, store
all the operands in locals and read from the locals, and then removing a
local.get is always safe to do, and nothing has been reordered:

(struct.new
  (A)
  (side effect) ;; this field will be removed
  (B)
)

=>

(local.set $a (A))
(local.set $t (side effect))
(local.set $b (B))
(struct.new
  (local.get $a)
  (local.get $b)
)

Later passes can remove unneeded local operations etc.

This is necessary before enabling this pass, as this corner case occurs on
j2wasm.