[Wasm GC] Enable GlobalTypeOptimization (#4305)
Now that all known issues with that pass are fixed, enable it by
default. This adds it in a place that seems to make sense on j2wasm,
but in general multiple cycles of optimization will be needed.

This adds a test showing that we run this pass and that it helps
ConstantFieldPropagation by running before it.