Effects.h: Fix RefAs (#4312)
We marked that as only trapping if the input as nullable. But ref.as_func
will trap if it isn't a func, for example.

We could in theory try to check if a trap is possible, like checking if the
input is already non-nullable or already a function, etc., but we have
optimization passes to get rid of RefAs when they are not needed
anyhow, so there is no point to duplicate that here.