[EH] Support try-delegate in EffectAnalyzer
This adds support for try-delegate in `EffectAnalyzer`. Without this
support, the expresion below has been incorrectly classified as "cannot
throw", because the previous code considered everything inside
`try`-`catch_all` as "cannot throw". This not the case when there is a
`delegate` that can bypass the `catch_all`.
```wasm
try $l0
  try
    try
      throw $e
    delegate $l0
  catch_all
  end
end