[EH] Handle nested pops after inlining
Inlining creates additional `block`s at inlined call sites, which can be
inside a `catch`. For example:
```wast
(try
  (do)
  (catch $tag
    (call $callee
      (pop i32)
    )
  )
)
```

After inlining, this becomes
```wast
(try
  (do)
  (catch $tag$0
    (block $__inlined_func$callee
      (local.set $0
        (pop i32) ;; Invalid!!
      )
    (nop)
    )
  )
)
```

Now the `pop` is nested in a `block`, which makes this invalid. This PR
runs `EHUtils::handleBlockNestedPops` at the end to assign the `pop` to
a local right after the `catch`, making the code valid again.