Add categories to --help text (#4421)
The general shape of the --help output is now:



========================
wasm-foo

Does the foo operation
========================

wasm-foo opts:
--------------

  --foo-bar   ..

Tool opts:
----------

..




The options are now in categories, with the more specific ones - most likely to be
wanted by the user - first. I think this makes the list a lot less confusing.

In particular, in wasm-opt all the opt passes are now in their own category.

Also add a script to make it easy to update the help tests.