[ctor-eval] Remove stack hacks (#4429)
Remove some hackish code for fastcomp's stack handling. The stack pointer arrives
in an imported global there. Upstream does not do this, so this code is completely
unneeded these days (and, frankly, kind of scary as I read it now... it modeled the
stack as separate memory from the heap...).

Remove the tests for this as well. I verified that there was nothing else in those
tests that we need to keep.