Escape \t as well as \n when writing JSON output.
As it happens, this doesn't (normally) break the resulting EM_ASM or
EM_JS strings because (IIUC) JS supports the tab literal inside of
strings as well as "\t".

However, it's better to preserve the original text so that it looks
the same in the JS file as it did in the original source.