SafeHeap: Avoid instrumenting any function called from the start function.
The `__wasm_memory_init` function is not always used as the start
function.  Sometimes it is just one of a sequence of functions that are
called from the start function.  In order to handle this case we can
instead avoid instrumenting the start function as well as any functions
call from it.

This fixes the SAFE_HEAP + USE_PTHREADS + MAIN_MODULE configuration
which is currently broken in emscripten.