[NFC] Templatize/generalize RuntimeExpressionRunner (#4477)
Add a base class for it, that is templated and can be extended in general
ways, and make callFunction templated on the runner to use as well.

This allows the interpreter's behavior to be customized in a way that we
couldn't so far. wasm-ctor-eval wants to use a special Runner when it
evals a function, so that it can track certain operations, which this will
enable.