Isorecursive type output ordering
Generally we try to order types by decreasing use count so that frequently used
types get smaller indices. For the equirecursive and nominal systems, there are
no contraints on the ordering of types, so we just have to sort them according
to their use counts. For the isorecursive type system, however, there are a
number of ordering constraints that have to be met for the type section to be
valid.

First, types in the same recursion group must be adjacent so they can be grouped
together. Next, groups must be ordered topologically. Finally, supertypes must
appear before their subtypes, even within a single group. Only after satisfying
these constraints can we additionally order by use count.

Update type ordering to produce a valid and optimal ordering. In principle it
would be possible to do less work for use cases that don't require the optimal
ordering, such as in type updating passes, but skipping unnecessary work
wouldn't improve the algorithmic complexity of the sorting, so don't bother for
now.